helper/pip_requirement.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.10
config==0.5.1
configobj==5.0.9
configparser==7.1.0
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.6
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.32.2
dill==0.3.8
distro==1.9.0
duckduckgo_search==7.2.1
edge-tts==7.0.0
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
fastjsonschema==2.20.0
feedparser==6.0.11
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.13.1
flatbuffers==24.12.23
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.2.0
future==1.0.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.13.1
kiwisolver==1.4.7
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.41.1
looseversion==1.3.0
lxml==5.3.1
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
modelscope==1.22.3
MouseInfo==0.1.3
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
mss==9.0.2
mtcnn==0.1.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.31.0
nbclassic==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
notebook==6.5.7
notebook_shim==0.2.4
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.2
openai==1.68.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optuna==4.2.1
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pickleshare==0.7.5
pillow==10.2.0
pip==24.2
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
playsound==1.2.2
pooch==1.8.2
primp==0.14.0
proglog==0.1.10
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
prov==2.0.1
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
pydub==0.25.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pykeyboard==0.1.5
PyMsgBox==1.0.9
PyMuPDF==1.24.11
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
python-pptx==1.0.2
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.2
pywin32==306
pywinpty==2.0.13
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
ruff==0.11.2
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
selenium==4.27.1
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools==75.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.0
soupsieve==2.6
soxr==0.3.7
SpeechRecognition==3.10.4
SQLAlchemy==2.0.39
srt==3.5.3
stack-data==0.6.3
starlette==0.44.0
statsmodels==0.14.1
streamlit==1.40.1
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-hub==0.16.1
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tf-keras==2.15.0
TgCrypto==1.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tkinterdnd2==0.4.3
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.4.1+cu118
torchaudio==2.4.1+cu118
torchvision==0.19.1+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
transformers==4.46.3
trio==0.27.0
trio-websocket==0.12.2
tweepy==4.14.0
typer==0.15.2
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
ultralytics==8.3.2
ultralytics-thop==2.0.8
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.33.0
volcengine-python-sdk==1.0.130
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wheel==0.44.0
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.1
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
